fix(lifi): address Phase 3 review caveats
- fix: pass force=false on first swap attempt; surface onchainos risk
  warnings then retry with force=true (user already confirmed via --confirm)
- fix: replace hardcoded 15s approval sleep with chain-specific wait
  (3s Base/Arb/Op, 6s Polygon/BSC/Avax, 20s Ethereum)
- fix: u64->u128 for native value parsing to avoid overflow on large txs

+++ skills/lifi/src/onchainos.rs
@@ -39,7 +39,7 @@ pub async fn wallet_contract_call(
     to: &str,
     input_data: &str,
     from: Option<&str>,
-    amt: Option<u64>, // wei value (for native ETH sends)
+    amt: Option<u128>, // wei value (for native ETH sends)
     dry_run: bool,
     force: bool,
 ) -> Result<Value> {
